Decompose tile_map_generator_land_seeds into part files (Refs #2199)
Split `tile_map_generator_land_seeds.dart` (914 lines) into a library
backed by four focused part files; the public `TileMapGenLandSeeds`
class is a thin facade delegating to free functions in the parts:

- `tile_map_generator_land_seeds_shared_part.dart`: Voronoi index range
  bookkeeping, per-cell best-continent / cell-entry assembly,
  continent-buffer join checks, and buffer-offset enumeration shared
  between placement and organic passes.
- `tile_map_generator_land_seeds_placement_part.dart`: seed-before-
  assignment placement (`_placeLandSeedsImpl`), final non-organic
  Voronoi assignment (`_assignLandByLandSeedsImpl`), and the Box-Muller
  Gaussian helper.
- `tile_map_generator_land_seeds_organic_part.dart`: organic seed
  candidate selection helpers, per-round `_placeOneOrganicSeed`,
  Voronoi-with-no-join, and the interleaved organic driver
  (`_placeLandSeedsOrganicImpl`).
- `tile_map_generator_land_seeds_coast_part.dart`: coastal tile
  registration, local-density scoring, per-continent single-step
  growth, and the thickness-first `_growCoastlines` driver.

Pure code-movement: signatures, control flow, and `_landSentinel`
usage are preserved; previously-instance methods that captured
`this.params` are now free functions taking `TileMapLandSeedParams`
explicitly. Public surface (the file path and `TileMapGenLandSeeds`
class API) is unchanged, so callers in `tile_map_generator.dart`, the
`colonizethis_map.dart` barrel, and tests need no edits.

Addresses AC: `tile_map_generator_land_seeds.dart` <= 300 lines
(now 73). Other ACs from #2199 are deferred to follow-up PRs.
